# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: TenderChain API
  version: 1.0.0
  description: |
    RESTful API for the TenderChain Blockchain-Based Government Tender Transparency System.
    Provides endpoints for tender management, sealed-bid procurement with commit-reveal scheme,
    audit trail querying, and user authentication.
  contact:
    name: TenderChain Team
    email: support@tenderchain.io
    url: https://tenderchain.io

servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.tenderchain.io
    description: Production server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via /api/auth/login or /api/auth/register

  schemas:
    Tender:
      type: object
      properties:
        id:
          type: integer
          description: Auto-incremented tender ID
        title:
          type: string
          description: Tender title
        description:
          type: string
          description: Detailed description
        ipfsDocumentHash:
          type: string
          description: IPFS hash of tender document
        submissionDeadline:
          type: string
          format: date-time
          description: Deadline for bid submissions
        revealDeadline:
          type: string
          format: date-time
          description: Deadline for bid reveals
        eligibilityCriteriaHash:
          type: string
          description: IPFS hash of eligibility criteria
        minimumBidAmount:
          type: string
          description: Minimum bid amount in wei
        estimatedProjectValue:
          type: string
          description: Estimated project value in wei
        status:
          type: string
          enum: [DRAFT, PUBLISHED, CLOSED, EVALUATION, AWARDED, CANCELLED]
        authority:
          type: string
          description: Wallet address of procuring authority
        createdAt:
          type: string
          format: date-time

    Bid:
      type: object
      properties:
        id:
          type: integer
        tenderId:
          type: integer
        bidder:
          type: string
          description: Bidder wallet address
        commitment:
          type: string
          description: Keccak256 commitment hash
        status:
          type: string
          enum: [COMMITTED, REVEALED, FORFEITED]
        createdAt:
          type: string
          format: date-time

    AuditLogEntry:
      type: object
      properties:
        id:
          type: integer
        timestamp:
          type: string
          format: date-time
        actor:
          type: string
          description: Wallet address of the actor
        actionType:
          type: string
          enum: [TENDER_CREATED, TENDER_PUBLISHED, TENDER_CANCELLED, BID_COMMITTED, BID_REVEALED, BID_FORFEITED, EVALUATION_STARTED, EVALUATION_COMPLETED, TENDER_AWARDED, APPEAL_FILED, APPEAL_RESOLVED]
        relatedEntityId:
          type: integer
        dataHash:
          type: string

    CommitmentRecord:
      type: object
      properties:
        valid:
          type: boolean
        tenderId:
          type: integer
        commitment:
          type: string
        timestamp:
          type: string
          format: date-time

    HealthStatus:
      type: object
      properties:
        api:
          type: string
          example: "ok"
        blockchain:
          type: string
          example: "ok"
        ipfs:
          type: string
          example: "ok"
        postgres:
          type: string
          example: "ok"

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string

paths:
  /api/auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account with email, password, role, and wallet address.
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, role, wallet, name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                role:
                  type: string
                  enum: [authority, contractor, auditor]
                wallet:
                  type: string
                  pattern: "^0x[a-fA-F0-9]{40}$"
                name:
                  type: string
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      refreshToken:
                        type: string
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/login:
    post:
      summary: Login with credentials
      description: Authenticates user with email and password, returns JWT token.
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/tenders:
    post:
      summary: Create a new tender
      description: Creates a new tender in DRAFT status. Requires AUTHORITY role.
      tags: [Tenders]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, ipfsDocumentHash, submissionDeadline, revealDeadline, minimumBidAmount]
              properties:
                title:
                  type: string
                description:
                  type: string
                ipfsDocumentHash:
                  type: string
                submissionDeadline:
                  type: string
                revealDeadline:
                  type: string
                minimumBidAmount:
                  type: string
                estimatedProjectValue:
                  type: string
      responses:
        "201":
          description: Tender created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Tender"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — role not authority

    get:
      summary: List tenders
      description: Returns paginated list of tenders. Public endpoint.
      tags: [Tenders]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated tender list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tender"
                  meta:
                    type: object

  /api/tenders/{id}:
    get:
      summary: Get tender by ID
      description: Returns full detail for a specific tender.
      tags: [Tenders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tender details
        "404":
          description: Tender not found

  /api/tenders/{id}/publish:
    post:
      summary: Publish a tender
      description: Changes tender status from DRAFT to PUBLISHED.
      tags: [Tenders]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tender published
        "400":
          description: Tender not in DRAFT status
        "404":
          description: Tender not found
        "409":
          description: Already published

  /api/bids/submit:
    post:
      summary: Submit an encrypted bid
      description: Submits a sealed bid with encrypted payload and commitment hash.
      tags: [Bids]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenderId, encryptedPayload, amount, technicalProposalHash]
              properties:
                tenderId:
                  type: integer
                encryptedPayload:
                  type: string
                amount:
                  type: string
                technicalProposalHash:
                  type: string
      responses:
        "201":
          description: Bid committed
        "401":
          description: Unauthorized
        "409":
          description: Duplicate bid

  /api/bids/verify/{tenderId}/{commitment}:
    get:
      summary: Verify a commitment
      description: Public endpoint to verify if a commitment exists for a tender.
      tags: [Bids]
      parameters:
        - name: tenderId
          in: path
          required: true
          schema:
            type: integer
        - name: commitment
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Commitment verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitmentRecord"

  /api/audit/logs:
    get:
      summary: Query audit logs
      description: Public endpoint to query immutable audit log entries with filtering.
      tags: [Audit]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: actionType
          in: query
          schema:
            type: string
        - name: entityId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditLogEntry"

  /api/audit/logs/{id}:
    get:
      summary: Get single audit log entry
      description: Returns a specific audit log entry by ID.
      tags: [Audit]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Log entry details
        "404":
          description: Entry not found

  /api/health:
    get:
      summary: Health check
      description: Returns the health status of all system components.
      tags: [System]
      responses:
        "200":
          description: All services healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/HealthStatus"